+++ pallets/xcm-core-buyer/Cargo.toml
@@ -15,6 +15,7 @@ dp-core = { workspace = true }
 log = { workspace = true }
 serde = { workspace = true, optional = true }
 tp-traits = { workspace = true }
+tp-xcm-core-buyer = { workspace = true }
 
 # Polkadot
 pallet-xcm = { workspace = true }
@@ -68,6 +69,7 @@ std = [
 	"staging-xcm/std",
 	"staging-xcm/std",
 	"tp-traits/std",
+	"tp-xcm-core-buyer/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
@@ -79,6 +81,7 @@ runtime-benchmarks = [
 	"pallet-xcm/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"tp-traits/runtime-benchmarks",
+	"tp-xcm-core-buyer/runtime-benchmarks",
 ]
 try-runtime = [
 	"frame-support/try-runtime",
