+++ primitives/xcm-core-buyer/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "tp-xcm-core-buyer"
+authors = { workspace = true }
+description = "XCM Core Buyer Primitives"
+edition = "2021"
+license = "GPL-3.0-only"
+version = "0.1.0"
+
+[lints]
+workspace = true
+
+[dependencies]
+frame-support = { workspace = true }
+parity-scale-codec = { workspace = true }
+scale-info = { workspace = true }
+sp-runtime = { workspace = true }
+tp-traits = { workspace = true }
+
+[features]
+default = [ "std" ]
+std = [
+	"frame-support/std",
+	"parity-scale-codec/std",
+	"scale-info/std",
+	"sp-runtime/std",
+	"tp-traits/std",
+]
+runtime-benchmarks = [
+	"frame-support/runtime-benchmarks",
+	"sp-runtime/runtime-benchmarks",
+	"tp-traits/runtime-benchmarks",
+]
